+++ resources/views/articles/form.blade.php
@@ -99,11 +99,17 @@
 	</div>
 </div>
 <div class="col-md-6">
-	<div class="form-group">
-		{!! Form::label('main_image', 'Obrázok') !!}
-		{!! Form::file('main_image') !!}
+	<div class="form-group ">
+		{!! Form::label('main_image', 'Obrázok') !!}	
+		{!! Form::file('main_image', array('class' => 'form-control', 'placeholder' => '#666666'))!!}
 		<p>obrazok bude automaticky zmenseny na sirku 1400px</p>
 		<p>šírka min: 1400px<br>formát: JPG (vysoka kompresia ~50-60%)</p>
+		@if (isset($article) && $article->getHeaderImage())
+			<div class="primary-image">
+				<b>Aktuálny obrázok:</b>
+				<img src="{{ $article->getHeaderImage() }}" class="img-responsive">
+			</div>
+		@endif
 	</div>
 </div>
 <div class="col-md-6">